+++ os/linux/cfg80211/cfg80211drv.c
@@ -797,7 +797,7 @@ BOOLEAN CFG80211DRV_StaKeyAdd(
 	{
 		RT_CMD_STA_IOCTL_SECURITY IoctlSec;
 
-		DBGPRINT(RT_DEBUG_TRACE, ("Set_WPAPSK_Proc ==> %d, %d, %d...\n", pKeyInfo->KeyId, pKeyInfo->KeyType, strlen(pKeyInfo->KeyBuf)));
+		DBGPRINT(RT_DEBUG_TRACE, ("Set_WPAPSK_Proc ==> %d, %d, %d...\n", pKeyInfo->KeyId, pKeyInfo->KeyType, (int) strlen(pKeyInfo->KeyBuf)));
 
 		IoctlSec.KeyIdx = pKeyInfo->KeyId;
 		IoctlSec.pData = pKeyInfo->KeyBuf;
